requirements.txt
yfinance
pandas
matplotlib
mplfinance
GoogleNews
textblob
tweepy
colorama
python-dotenv